+++ pakyow-core/spec/unit/generator_spec.rb
@@ -7,6 +7,18 @@ RSpec.describe Pakyow::Generator do
     File.expand_path("../support/generator", __FILE__)
   end
 
+  let :destination_path do
+    File.expand_path("../support/generated", __FILE__)
+  end
+
+  let :generator_class do
+    described_class
+  end
+
+  let :instance do
+    generator_class.new(source_path)
+  end
+
   describe "#initialize" do
     it "initializes with a path" do
       expect {
@@ -16,25 +28,17 @@ RSpec.describe Pakyow::Generator do
   end
 
   describe "#generate" do
-    let :destination_path do
-      File.expand_path("../support/generated", __FILE__)
-    end
-
     let :options do
       {
         foo: "bar"
       }
     end
 
-    let :instance do
-      described_class.new(source_path)
-    end
-
     it "generates each file" do
       expect(instance.files.length).to eq(5)
 
       instance.files.each do |file|
-        expect(file).to receive(:generate).with(Pathname.new(destination_path), options)
+        expect(file).to receive(:generate).with(Pathname.new(destination_path), context: instance_of(described_class))
       end
 
       instance.generate(destination_path, options)
@@ -86,10 +90,6 @@ RSpec.describe Pakyow::Generator do
   end
 
   describe "#run" do
-    let :instance do
-      described_class.new(source_path)
-    end
-
     let :runner_double do
       instance_double(Pakyow::Support::CLI::Runner)
     end
@@ -106,50 +106,29 @@ RSpec.describe Pakyow::Generator do
       instance.run("ls", message: "foo")
     end
   end
-end
 
-RSpec.describe Pakyow::Generator::File do
-  let :source_path do
-    File.expand_path("../support/generator", __FILE__)
-  end
+  describe "generated files" do
+    let(:source_path) {
+      File.expand_path("../support/generator/test.txt", __FILE__)
+    }
 
-  let :file_path do
-    "test.txt"
-  end
+    let(:generated_file_path) {
+      "test.txt"
+    }
 
-  let :destination_path do
-    File.expand_path("../support/generated", __FILE__)
-  end
-
-  let :instance do
-    generator_class.new(File.join(source_path, file_path), source_path)
-  end
-
-  let :generator_class do
-    described_class
-  end
-
-  describe "#initialize" do
-    it "initializes with a path and source path" do
-      expect {
-        described_class.new(File.join(source_path, file_path), source_path)
-      }.to_not raise_error
-    end
-  end
-
-  describe "#generate" do
     before do
       cleanup
-      instance.generate(destination_path, options)
+
+      instance.generate(destination_path, **options)
     end
 
     after do
       cleanup
     end
 
-    let :options do
+    let(:options) {
       {}
-    end
+    }
 
     def cleanup
       if File.exist?(destination_path)
@@ -158,26 +137,30 @@ RSpec.describe Pakyow::Generator::File do
     end
 
     it "generates the file at the destination" do
-      expect(File.exist?(File.join(destination_path, file_path))).to be(true)
+      expect(File.exist?(File.join(destination_path, generated_file_path))).to be(true)
     end
 
     describe "the generated file" do
       it "contains the correct content" do
-        expect(File.read(File.join(destination_path, file_path))).to eq("this is a plain file test\n")
+        expect(File.read(File.join(destination_path, generated_file_path))).to eq("this is a plain file test\n")
       end
     end
 
     context "file is an erb template" do
-      let :file_path do
-        "test.txt.erb"
-      end
+      let(:source_path) {
+        File.expand_path("../support/generator/test.txt.erb", __FILE__)
+      }
+
+      let(:generated_file_path) {
+        "test.txt"
+      }
 
       it "renames the generated file" do
-        expect(File.exist?(File.join(destination_path, "test.txt"))).to be(true)
+        expect(File.exist?(File.join(destination_path, generated_file_path))).to be(true)
       end
 
       it "evaluates the template" do
-        expect(File.read(File.join(destination_path, "test.txt")).strip).to eq("2")
+        expect(File.read(File.join(destination_path, generated_file_path)).strip).to eq("2")
       end
 
       context "options are passed" do
@@ -195,39 +178,47 @@ RSpec.describe Pakyow::Generator::File do
     end
 
     context "file is a dotfile" do
-      let :file_path do
-        "%dot%gitignore"
-      end
+      let(:source_path) {
+        File.expand_path("../support/generator/%dot%gitignore", __FILE__)
+      }
+
+      let(:generated_file_path) {
+        ".gitignore"
+      }
 
       it "renames the generated file" do
-        expect(File.exist?(File.join(destination_path, ".gitignore"))).to be(true)
+        expect(File.exist?(File.join(destination_path, generated_file_path))).to be(true)
       end
 
       describe "the generated file" do
         it "contains the correct content" do
-          expect(File.read(File.join(destination_path, ".gitignore")).strip).to eq("dotfile_test")
+          expect(File.read(File.join(destination_path, generated_file_path)).strip).to eq("dotfile_test")
         end
       end
     end
 
     context "file is a keep file" do
-      let :file_path do
-        "test-keep/keep"
-      end
+      let(:source_path) {
+        File.expand_path("../support/generator/test-keep", __FILE__)
+      }
 
       it "generates the enclosing directory" do
-        expect(File.directory?(File.join(destination_path, "test-keep"))).to be(true)
+        expect(File.directory?(File.join(destination_path, "folder"))).to be(true)
       end
 
       it "does not generate the keep file" do
-        expect(File.exist?(File.join(destination_path, "test-keep/keep"))).to be(false)
+        expect(File.exist?(File.join(destination_path, "folder/keep"))).to be(false)
       end
     end
 
     context "with a subclass" do
-      let :file_path do
-        "test-methods.txt.erb"
-      end
+      let(:source_path) {
+        File.expand_path("../support/generator/test-methods.txt.erb", __FILE__)
+      }
+
+      let(:generated_file_path) {
+        "test-methods.txt"
+      }
 
       let :generator_class do
         Class.new(described_class) do
@@ -238,7 +229,7 @@ RSpec.describe Pakyow::Generator::File do
       end
 
       it "exposes content through methods" do
-        expect(File.read(File.join(destination_path, "test-methods.txt")).strip).to eq("bar")
+        expect(File.read(File.join(destination_path, generated_file_path)).strip).to eq("bar")
       end
     end
   end